Add deprecation notice and clean out repository
We've decided to deprecate our older "sample application", and point
customers to the other "demo application".

I've updated the README, and also removed the older files so as to
avoid confusion, but in case they are ever needed -- git remembers.

Next step after this is marking the repository as "archived".